Fixes to get compilation working in VS.
git-svn-id: https://svn.code.sf.net/p/kaldi/code/trunk@433 5e6a8d80-dfce-4ca6-a32a-6e07a63d50c8

+++ src/nnet/nnet-rnnlm.cc
@@ -205,7 +205,7 @@ BaseFloat Rnnlm::Score(const std::vector<int32>& seq, const VectorBase<BaseFloat
   Propagate(in,&out);
 
   BaseFloat llk = 0.0;
-  for(int32 i=0; i<seq.size(); i++) {
+  for(size_t i=0; i<seq.size(); i++) {
     llk += log(out(i,seq[i]-1));
   }
